+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Every release, along with the migration instructions, is documented on the Githu
 ## UNRELEASED
 
 - Add `DashboardPlugin({ includeAssets: [ "stringPrefix", /regexObj/ ] })` Webpack plugin filtering option.
-- Add `webpack-dashboard --include-assets stringPrefix1 -a stringPrefix1` CLI filtering option.
+- Add `webpack-dashboard --include-assets stringPrefix1 -a stringPrefix2` CLI filtering option.
 - Change `"mode"` SocketIO event to `"options"` as it now passes both `minimal` and `includeAssets` from CLI to the Webpack plugin.
 - Fix unit tests that incorrectly relied on `.complete()` for `most` observables.
 - Add additional `examples` fixture for development.